Fixed errors in put and delete for worker, as well as long to int in id machine

+++ src/main/java/com/example/wokers_machines_l2/repository/MachineRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MachineRepository extends JpaRepository<Machine, Long> {
+public interface MachineRepository extends JpaRepository<Machine, Integer> {
 }
